LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/alignment/flex-align-baseline-003-expected.txt

line1
line2
line1
line2
line1
line2
line1
line2
line1
line2
line1
line2

PASS #target > div 1
FAIL #target > div 2 assert_equals:
<div data-offset-x="105">line1<br>line2</div>
offsetLeft expected 105 but got 0
FAIL #target > div 3 assert_equals:
<div data-offset-x="126">line1<br>line2</div>
offsetLeft expected 126 but got 21
FAIL #target > div 4 assert_equals:
<div data-offset-x="20">line1<br>line2</div>
offsetLeft expected 20 but got 125
PASS #target > div 5
PASS #target > div 6
